various

+++ php/remove.php
@@ -13,6 +13,12 @@ if (!$stmt->execute()) { panic('SQL Error!'); }
 if ($stmt->get_result()->fetch_row()) {
   panic('You cannot delete a post that already has content!');
 }
+$stmt = mysqli()->prepare('SELECT `image` FROM `post_free` WHERE `post_id` = ?');
+$stmt->bind_param('i', $legal_post_id);
+if (!$stmt->execute()) { panic('SQL Error!'); }
+if ($stmt->get_result()->fetch_row()[0]) {
+  panic('You cannot delete a post that already has an image!');
+}
 $stmt = mysqli()->prepare('DELETE FROM `post_free` WHERE `post_id` = ?');
 $stmt->bind_param('i', $legal_post_id);
 if (!$stmt->execute()) { panic('SQL Error!'); }
